+++ lib/explicit/trinkets/include/enumerable/each_with_hash.rb
@@ -1,0 +1,14 @@
+module Trinkets
+
+  module Enumerable
+    module WithHash
+      def each_with_hash(&) = each_with_object({}, &)
+    end
+  end
+
+  module Enumerator
+    module WithHash
+      def with_hash(&) = each_with_hash(&)
+    end
+  end
+end