+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 
+### [0.3.0](https://github.com/SilverPhoenix99/trinkets/tree/v0.3.0)
+* Restructured folders to default to `/refine/`.
+* Added `Enumerable#each_with_hash`/`Enumerator#with_hash`.
+
 ### [0.2.0](https://github.com/SilverPhoenix99/trinkets/tree/v0.2.0)
 * Added `attr: :none` as an option for `Class::init`.